Refactoring of lib directory
