+++ lib/line.js
