+++ lib/novel_reviews.js
