+++ lib/novel_writer.js
