+++ scripts/scraping/kakuyomu_reviews_export.js
@@ -1,7 +1,7 @@
 const puppeteer = require('puppeteer');
 
-const NovelReviews = require('../lib/novel_reviews');
-const NovelWriter = require('../lib/novel_writer');
+const NovelReviews = require('../../lib/novel_reviews');
+const NovelWriter = require('../../lib/novel_writer');
 
 
 (async () => {
